requirements.txt
decorator==4.4.2
dlib==19.7.0
imageio==2.16.0
imutils==0.5.4
networkx==2.5.1
numpy==1.19.4
opencv-python==4.4.0.46
pandas==1.1.4
Pillow==8.0.1
pygame==2.1.2
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.1.1
scikit-image==0.15.0
scipy==1.5.4
six==1.16.0
